requirements.txt
openai-whisper
torch
torchaudio
ffmpeg-python
flask
flask-sqlalchemy
tqdm